Seem to be working but still issue with gradTest
It seems to be related to the fact that |gradJ|^2
is much smaller than J that the residual drop to
0.  Worse at times it gives an absurd negative
number.
The problem is not in the gradTest routine.  I've
tested it with a simple cost function and it work
perfectly altough if the former problem arises,
say f=x^2+999999999999999, it still gives 0.