JTerm for dynamic obs with preconditionning
There is still the issue that the gradient
test seems to indicate only float32 precision.
It would mean an implicit conversion somewhere
down the chain....
In the Fortran/Python interface (for pyKdV
tries)?
In scipy.optimize.fmin_bfgs?

So till now PrecondJTWObsJTerm does not work very
well, I think it's due to that.