Work around the nan gradient problem
The problem is still there:
After imposing a maximum on the norm of the
gradient, the minimizer still ends in state with
huge amplitude returning a nan gradient.
So the minimizer stops and return the huge valued
state with nan gradient.
When that happen, we want the previous state as
analysis.